% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html-helpers.R
\name{add_badge_color}
\alias{add_badge_color}
\title{Add badge color}
\usage{
add_badge_color(add_color, add_label, alpha_lvl)
}
\arguments{
\item{add_color}{A color to add to the badge}

\item{add_label}{The label to add to the badge}

\item{alpha_lvl}{The alpha level}
}
\value{
HTML character
}
\description{
Add badge color
}
