% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_plt_conf_int.R
\name{add_ci_plot}
\alias{add_ci_plot}
\title{Add a confidence interval plot inside a specific row}
\usage{
add_ci_plot(data_in, pal_vals, width, ext_range, text_args)
}
\arguments{
\item{data_in}{A dataframe of length 1}

\item{pal_vals}{A length 4 palette to be used for coloring points, segments and text}

\item{width}{Width of the output plot in \code{'mm'}}

\item{ext_range}{A length two vector of the full range across all values}

\item{text_args}{A list of optional text arguments passed to \code{scales::label_number_si()}}
}
\value{
SVG/HTML
}
\description{
Add a confidence interval plot inside a specific row
}
