% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_plt_percentile_dot.R
\name{add_pcttile_plot}
\alias{add_pcttile_plot}
\title{Create a dot plot from 0 to 100}
\usage{
add_pcttile_plot(data, palette, add_label, width)
}
\arguments{
\item{data}{The single value that will be used to plot the point.}

\item{palette}{A length 3 palette, used to highlight high/med/low}

\item{add_label}{A logical indicating whether to add the label or note. This will only be added if it is the first or last row.}

\item{width}{A numeric indicating the}
}
\value{
gt table
}
\description{
Create a dot plot from 0 to 100
}
