% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_plt_point.R
\name{add_point_plot}
\alias{add_point_plot}
\title{Create a dot plot from any range - add_point_plot}
\usage{
add_point_plot(data, palette, add_label, width, vals_range, accuracy)
}
\arguments{
\item{data}{The single value that will be used to plot the point.}

\item{palette}{A length 3 palette, used to highlight high/med/low}

\item{add_label}{A logical indicating whether to add the label or note. This will only be added if it is the first or last row.}

\item{width}{A numeric indicating the}
}
\value{
gt table
}
\description{
Create a dot plot from any range - add_point_plot
}
