% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_summary_table.R
\name{create_sum_table}
\alias{create_sum_table}
\title{Create a summary table from a dataframe}
\usage{
create_sum_table(df)
}
\arguments{
\item{df}{a dataframe or tibble}
}
\value{
A summary dataframe as a tibble
}
\description{
Create a summary table from a dataframe
}
\examples{
\dontrun{
create_sum_table(iris)
#>   # A tibble: 5 × 7
#>   type    name         value       n_missing  Mean Median     SD
#>   <chr>   <chr>        <list>          <dbl> <dbl>  <dbl>  <dbl>
#> 1 numeric Sepal.Length <dbl [150]>         0  5.84   5.8   0.828
#> 2 numeric Sepal.Width  <dbl [150]>         0  3.06   3     0.436
#> 3 numeric Petal.Length <dbl [150]>         0  3.76   4.35  1.77
#> 4 numeric Petal.Width  <dbl [150]>         0  1.20   1.3   0.762
#> 5 factor  Species      <fct [150]>         0  NA     NA    NA
}

}
