% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html-helpers.R
\name{gt_hyperlink}
\alias{gt_hyperlink}
\title{Add a basic hyperlink in a gt table}
\usage{
gt_hyperlink(text, url)
}
\arguments{
\item{text}{The text displayed for the hyperlink}

\item{url}{The url for the hyperlink}
}
\value{
HTML text
}
\description{
A lightweight helper to add a hyperlink, can be used throughout
a \code{gt} table.
}
