% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_pct_bar.R
\name{gt_plt_bar_stack}
\alias{gt_plt_bar_stack}
\title{Add a percent stacked barchart in place of existing data.}
\usage{
gt_plt_bar_stack(
  gt_object,
  column = NULL,
  palette = c("#ff4343", "#bfbfbf", "#0a1c2b"),
  labels = c("Group 1", "Group 2", "Group 3"),
  width = 70
)
}
\arguments{
\item{gt_object}{An existing gt table object of class \code{gt_tbl}}

\item{column}{The column wherein the percent stacked barchart should replace existing data. Note that the data \emph{must} be represented as a list of numeric values ahead of time.}

\item{palette}{A color palette of length 2 or 3, represented either by hex colors (\verb{"#ff4343}) or named colors (\code{"red"}).}

\item{labels}{A vector of strings of length 2 or 3, representing the labels for the bar chart, will be colored according to the palette as well.}

\item{width}{An integer representing the width of the bar chart in pixels.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
The \code{gt_plt_bar_stack} function takes an existing \code{gt_tbl} object and
converts the existing values into a percent stacked barchart. The bar chart
will represent either 2 or 3 user-specified values per row, and requires
a list column ahead of time. The palette and labels need to be equal length.
}
\section{Function ID}{

1-5
}

\seealso{
Other Plotting: 
\code{\link{gt_bar_plot}()},
\code{\link{gt_kable_sparkline}()},
\code{\link{gt_plt_winloss}()},
\code{\link{gt_sparkline}()}
}
\concept{Plotting}
