% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_summary_table.R
\name{gt_plt_summary}
\alias{gt_plt_summary}
\title{Create a summary table from a dataframe}
\usage{
gt_plt_summary(df, title = NULL)
}
\arguments{
\item{df}{a dataframe or tibble}

\item{title}{a character string to be used in the table title}
}
\value{
a gt table
}
\description{
Create a summary table from a dataframe with inline
histograms or area bar charts. Inspired by the Observable team and
the observablehq/SummaryTable function: https://observablehq.com/d/d8d2929832202050
}
