% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_render_reprex.R
\name{gt_render_reprex}
\alias{gt_render_reprex}
\title{Helper for gt reprex screenshots}
\usage{
gt_render_reprex(gt_object, imgur = FALSE)
}
\arguments{
\item{gt_object}{an existing gt table}

\item{imgur}{Logical indicates whether to print a markdown link to imgur upload or not}
}
\value{
a png image
}
\description{
Take a gt pipeline or object and print it within reprex
}
