% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_reprex_image.R
\name{gt_reprex_image}
\alias{gt_reprex_image}
\title{Render 'gt' Table to Temporary png File}
\usage{
gt_reprex_image(gt_object)
}
\arguments{
\item{gt_object}{An object of class \code{gt_tbl} usually created by \code{\link[gt:gt]{gt::gt()}}}
}
\value{
a png image
}
\description{
Take a gt pipeline or object and print it as an image within
a reprex
}
\details{
Saves a gt table to a temporary png image file and uses
\code{knitr::include_graphics()} to render tables in reproducible examples
like \code{reprex::reprex()} where the HTML is not transferrable to GitHub.
}
