% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_text_img.R
\name{gt_text_img}
\alias{gt_text_img}
\title{Add text and an image to the left or right of it}
\usage{
gt_text_img(text, url, height = 30, left = FALSE)
}
\arguments{
\item{text}{A text string to be added to the cell.}

\item{url}{A url that resolves to an image file.}

\item{height}{The absolute height (px) of the image in the table cell.}

\item{left}{A logical TRUE/FALSE indicating if text should be on the left (TRUE) or right (FALSE)}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
The \code{gt_text_img} function takes an existing \code{gt_tbl} object and
adds some user specified text and an image url to a specific cell. This is a
wrapper raw HTML strings and \code{gt::web_image()}. Intended to be used inside
the header of a table via \code{gt::tab_header()}.
}
\section{Function ID}{

1-7
}

\seealso{
Other Utilities: 
\code{\link{fmt_pad_num}()},
\code{\link{fmt_symbol_first}()},
\code{\link{gt_bar_plot}()},
\code{\link{gt_img_rows}()},
\code{\link{gt_kable_sparkline}()},
\code{\link{gt_merge_stack}()},
\code{\link{gt_plt_winloss}()},
\code{\link{pad_fn}()}
}
\concept{Utilities}
