% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_img_circle.R
\name{img_circle}
\alias{img_circle}
\title{Create a circular border around a}
\usage{
img_circle(value, height, border_color, border_weight)
}
\arguments{
\item{value}{The source image}

\item{height}{The height in pixels of the circle}

\item{border_color}{A string indicating the color of the border}

\item{border_weight}{The weight of the border in pixels}
}
\value{
HTML
}
\description{
Create a circular border around a
}
