% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_img_circle.R
\name{img_square}
\alias{img_square}
\title{Create a square colored border around an image}
\usage{
img_square(value, height, width, border_color, border_weight)
}
\arguments{
\item{value}{The source image}

\item{height}{The height in pixels of the circle}

\item{border_color}{A string indicating the color of the border}

\item{border_weight}{The weight of the border in pixels}
}
\value{
HTML
}
\description{
Create a square colored border around an image
}
