% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_row_id.R
\name{last_row_id}
\alias{last_row_id}
\title{Get last row id/index even by group}
\usage{
last_row_id(gt_object)
}
\arguments{
\item{gt_object}{An existing gt table object of class \code{gt_tbl}}
}
\description{
Get last row id/index even by group
}
