% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_summary_table.R
\name{plot_data}
\alias{plot_data}
\title{Create inline plots for a summary table}
\usage{
plot_data(col, ...)
}
\arguments{
\item{col}{The column of data to be used for plotting}

\item{...}{additional arguments passed to scales::label_number()}
}
\value{
svg text encoded as HTML
}
\description{
Create inline plots for a summary table
}
