% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html-helpers.R
\name{with_tooltip}
\alias{with_tooltip}
\title{A helper to add basic tooltip inside a gt table}
\usage{
with_tooltip(label, tooltip)
}
\arguments{
\item{label}{The label for the item with a tooltip}

\item{tooltip}{The text based tooltip for the item}
}
\value{
HTML text
}
\description{
This is a lightweight helper to add tooltip, typically to be
used within \code{gt::cols_label()}.
}
