% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html-helpers.R
\name{gt_label_details}
\alias{gt_label_details}
\title{Add a simple table with column names and matching labels}
\usage{
gt_label_details(label, content)
}
\arguments{
\item{label}{A string representing the label for the details expansion section.}

\item{content}{A named list}
}
\value{
HTML text
}
\description{
Add a simple table with column names and matching labels
}
