% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_summary_table.R
\name{gt_plt_summary}
\alias{gt_plt_summary}
\title{Create a summary table from a dataframe}
\usage{
gt_plt_summary(df, title = NULL)
}
\arguments{
\item{df}{a dataframe or tibble}

\item{title}{a character string to be used in the table title}
}
\value{
a gt table
}
\description{
Create a summary table from a dataframe with inline
histograms or area bar charts. Inspired by the Observable team and
the observablehq/SummaryTable function: https://observablehq.com/d/d8d2929832202050
}
\section{Examples}{


Create a summary table from a \code{data.frame} or \code{tibble}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{gt_plt_summary(datasets::ChickWeight)
}\if{html}{\out{</div>}}

\if{html}{\out{
<img src="https://raw.githubusercontent.com/jthomasmock/gtExtras/master/images/gt_plt_summary-chicks.png" style="width:100\%;">
}}
}

