% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_plotting.R
\name{gt_plt_winloss}
\alias{gt_plt_winloss}
\title{winloss_plot
Add win loss point plot into rows of a \code{gt} table}
\usage{
gt_plt_winloss(
  gt_object,
  column,
  colors = c("#D50A0A", "#013369", "gray"),
  max_wins = 17
)
}
\arguments{
\item{gt_object}{An existing gt table object of class \code{gt_tbl}}

\item{column}{The column wherein the winloss plot should replace existing data. Note that the data \emph{must} be represented as a list of numeric values ahead of time.}

\item{colors}{A character string of length 3, specifying the colors for loss, win, tie in that exact order.}

\item{max_wins}{An integer indicating the max possible wins, this will be used to add padding if the total wins/losses observed is less than the max. This is useful for mid-season reporting.}

\item{height}{Number representing the vertical height of the plot in pixels. Defaults to 45 px.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
The \code{gt_plt_winloss} function takes an existing \code{gt_tbl} object and
adds squares of a specific color and vertical position based on wins/losses.
It is a wrapper around \code{gt::text_transform()}.
}
\section{Function ID}{

1-5
}

\examples{

}
\seealso{
Other Utilities: 
\code{\link{fmt_symbol_first}()},
\code{\link{gt_bar_plot}()},
\code{\link{gt_img_rows}()},
\code{\link{gt_kable_sparkline}()},
\code{\link{gt_merge_stack}()},
\code{\link{pad_fn}()}
}
\concept{Utilities}
