% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtExtras-package.R
\docType{package}
\name{gtExtras-package}
\alias{gtExtras}
\alias{gtExtras-package}
\title{gtExtras: A Collection of Helper Functions for the gt Package}
\description{
Provides utility functions and common use-cases applied via the gt package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jthomasmock/gtExtras}
  \item Report bugs at \url{https://github.com/jthomasmock/gtExtras/issues}
}

}
\keyword{internal}
