% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtExtras-package.R
\docType{package}
\name{gtExtras-package}
\alias{gtExtras}
\alias{gtExtras-package}
\title{gtExtras: A Collection of Helper Functions for the gt Package}
\description{
Provides utility functions and common use-cases applied via the gt package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jthomasmock/gtExtras}
  \item Report bugs at \url{https://github.com/jthomasmock/gtExtras/issues}
}

}
\author{
\strong{Maintainer}: Thomas Mock \email{j.thomasmock@gmail.com}

Other contributors:
\itemize{
  \item Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID}) [contributor]
}

}
\keyword{internal}
