% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtExtras-package.R
\docType{package}
\name{gtExtras-package}
\alias{gtExtras}
\alias{gtExtras-package}
\title{gtExtras: Extending 'gt' for Beautiful HTML Tables}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides additional functions for creating beautiful tables with 'gt'. The functions are generally wrappers around boilerplate or adding opinionated niche capabilities and helpers functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jthomasmock/gtExtras}
  \item \url{https://jthomasmock.github.io/gtExtras/}
  \item Report bugs at \url{https://github.com/jthomasmock/gtExtras/issues}
}

}
\author{
\strong{Maintainer}: Thomas Mock \email{j.thomasmock@gmail.com} [copyright holder]

Other contributors:
\itemize{
  \item Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID}) [contributor]
}

}
\keyword{internal}
